# speed up `generate(type = "draw")`